refactor: break session.ts circular dependencies — import types from session-types.ts
session.ts had 7 circular dependency cycles caused by modules importing
types (SessionInfo, UIState, SessionState) from session.ts while session.ts
imported those modules' values. This blocked #4227 (server.ts extraction)
since the circular graph made it unsafe to move session.ts exports.

Fix: redirect type-only imports to session-types.ts, which was created
for this purpose (#4228) but was not consistently used.

Cycles broken: 11 → 5 (7 session.ts cycles eliminated)
Remaining 5 cycles are in unrelated modules (ACP, metrics, pipeline, telegram).

Files changed: 8 (one-line import redirect each)
- services/session/permissions.ts — SessionInfo
- services/session/persistence.ts — SessionInfo, SessionState
- session-transcripts.ts — SessionInfo, UIState
- session-discovery.ts — SessionInfo
- services/state/state-store.ts — SessionInfo, SessionState
- mcp/client.ts — SessionInfo
- services/interfaces.ts — SessionInfo
- stall-detector.ts — SessionInfo, UIState

Unblocks #4227 (server.ts route extraction).
